+++ e2e/fixture-projects/node/preprocess.sh
@@ -16,15 +16,4 @@ fi
 echo "Copying package.json from $SOURCE_DIR to $TARGET_DIR"
 cp "$SOURCE_DIR/package.json" "$TARGET_DIR/package.json"
 
-cd "$TARGET_DIR"
-
-echo "Adding missing dependencies in $TARGET_DIR"
-pnpm add @matterlabs/hardhat-zksync-node@1.0.3
-pnpm add ethers@6.0.0
-
-# Use sed to remove the "workspace:" prefix because pnpm is not properly installing zksync-node plugin
-sed -i 's/"workspace:1\.0\.3"/"1.0.3"/' package.json
-
-cd -
-
 echo "Pre-processing complete."