+++ packages/hardhat-zksync-upgradable/src/task-names.ts
@@ -1,0 +1,4 @@
+export const TASK_DEPLOY_ZKSYNC_BEACON = 'deploy-zksync:beacon';
+export const TASK_UPGRADE_ZKSYNC_BEACON = 'upgrade-zksync:beacon';
+export const TASK_DEPLOY_ZKSYNC_PROXY = 'deploy-zksync:proxy';
+export const TASK_UPGRADE_ZKSYNC_PROXY = 'upgrade-zksync:proxy';